// src/app/model/ministry/ministry-cust-details.model.ts
export interface MinistryCustomerDetails {
    customerNo: string;
    customerNameBN: string | null;
    customerNameEN: string | null;
    divisionCode: string | null;
    districtCode: string | null;
    upazilaCode: string | null;
    ministryCode: string | null;
    departmentCode: string | null;
    citycorporationCode: string | null;
    pouroshovaCode: string | null;
    unionParishadCode: string | null;
    religiousCode: string | null;
    nonBengaliCampCode: string | null;
    isPolice: string | null;
    zoneCode: string | null;
    circleCode: string | null;
    dbCode: string | null;
    locationCode: string | null;
}